+++ Sources/SwanKit/Types.swift
@@ -1,3 +1,5 @@
+// Primitive types
+
 /// Byte type.
 public typealias SWKByte   = Int8
 
@@ -15,3 +17,45 @@ public typealias SWKFloat  = Float
 
 /// Double type.
 public typealias SWKDouble = Double
+
+
+// Storage types
+
+/// Byte storage.
+public typealias SWKByteStorage  = SWKStorage<SWKByte>
+
+/// Short storage.
+public typealias SWKShortStorage = SWKStorage<SWKShort>
+
+/// Int storage.
+public typealias SWKIntStorage   = SWKStorage<SWKInt>
+
+/// Long storage.
+public typealias SWKLongStorage  = SWKStorage<SWKLong>
+
+/// Float storage.
+public typealias SWKFloatStorage  = SWKStorage<SWKFloat>
+
+/// Double storage.
+public typealias SWKDoubleStorage = SWKStorage<SWKDouble>
+
+
+// Tensor types
+
+/// Byte tensor.
+public typealias SWKByteTensor   = SWKTensor<SWKByte>
+
+/// Short tensor.
+public typealias SWKShortTensor  = SWKTensor<SWKShort>
+
+/// Int tensor.
+public typealias SWKIntTensor    = SWKTensor<SWKInt>
+
+/// Long tensor.
+public typealias SWKLongTensor   = SWKTensor<SWKLong>
+
+/// Float tensor.
+public typealias SWKFloatTensor  = SWKTensor<SWKFloat>
+
+/// Double tensor.
+public typealias SWKDoubleTensor = SWKTensor<SWKDouble>
