+++ Tests/SwanKitTests/SwanKitTests.swift
@@ -5,10 +5,10 @@ class SwanKitTests: XCTestCase {
   func testAllSpecs() {
     SSpec.run {
       spec_SWKSize()
-      spec_SWKDevise()
-      spec_Dimensionality()
+      spec_SWKStride()
       spec_SWKStorage()
       spec_SWKTensor()
+      spec_SWKDevise()
     }
 
     XCTAssert(SSpec.hasErrors == false)
